requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
lxml==5.3.2
matplotlib==3.10.1
narwhals==1.35.0
numpy==2.2.4
openai==1.74.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
wheel==0.45.1
